+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+## [0.8.0] - 2018-12-29
+
 ### Added
 
 - Constructor for standard logger for errors
@@ -117,7 +120,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Initial release
 
 
-[Unreleased]: https://github.com/goph/logur/compare/v0.7.1...HEAD
+[Unreleased]: https://github.com/goph/logur/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/goph/logur/compare/v0.7.1...v0.8.0
 [0.7.1]: https://github.com/goph/logur/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/goph/logur/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/goph/logur/compare/v0.5.0...v0.6.0